requirements.txt
Flask==0.11.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.1
PyJWT==1.4.0
SQLAlchemy==1.0.14
Werkzeug==0.11.9
